+++ mii/utils.py
@@ -8,7 +8,6 @@ import logging
 import importlib
 import torch
 import mii
-from transformers import Conversation
 from huggingface_hub import HfApi
 
 from mii.constants import (CONVERSATIONAL_NAME,
@@ -219,29 +218,6 @@ def get_provider_name(model_name, task):
     else:
         provider = mii.constants.MODEL_PROVIDER_NAME_HF
     return provider
-
-
-def create_conversation(request, **kwargs):
-    if isinstance(request, dict):
-        assert 'text' in request and 'past_user_inputs' in request and 'generated_responses' in request, "Conversation requires 'text', 'past_user_inputs', and 'generated_responses' keys"
-        text = request['text']
-        conversation_id = request[
-            'conversation_id'] if 'conversation_id' in request else None
-        past_user_inputs = request['past_user_inputs']
-        generated_responses = request['generated_responses']
-
-    else:
-        text = getattr(request, 'text')
-        conversation_id = getattr(request, 'conversation_id')
-        past_user_inputs = getattr(request, 'past_user_inputs')
-        generated_responses = getattr(request, 'generated_responses')
-
-    conv = Conversation(text=text,
-                        conversation_id=conversation_id,
-                        past_user_inputs=past_user_inputs,
-                        generated_responses=generated_responses,
-                        **kwargs)
-    return conv
 
 
 log_levels = {
